requirements.txt
django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
#django-cors-headers==4.3.1